Tweaking the lightning display
